feat: User主键规范化 + 注册登录模块清理
1. User实体: username(主键) → id(Long自增主键), 所有外键引用id
   - Rental/Order/Feedback 的 @JoinColumn 统一引用 User.id
   - UserRepository 主键类型 String → Long

2. 注册登录清理:
   - 删除 UserService 死方法: login(), registerAdmin(), getUserById(Integer), deleteUser(Integer)
   - RuntimeException → UserRegistrationException
   - DTO 加 @Valid 校验注解, GlobalExceptionHandler 加 MethodArgumentNotValidException 处理
   - LoginResponse 新增 id 字段, 前端此前取值为 undefined
   - AuthController 删除冗余 @CrossOrigin
   - UserDTO.id Integer → Long

+++ src/main/java/scooterrent/dto/UserDTO.java
@@ -5,10 +5,8 @@ import java.time.LocalDateTime;
 
 @Data
 public class UserDTO {
-    // ID字段保留但作为可选字段，用于兼容现有前端代码
-    private Integer id;
-    
-    // 用户名现在作为主键
+    private Long id;
+
     private String username;
     private String email;
     private String phone;
